+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- fix: Wait for 1 minute for transaction to be processed in
+  get_transaction_receipt rpc
 - ci: Fix starknet foundry sncast not found
 - fix: Ensure transaction checks are compatible with starknet-rs
 - ci: Run Starknet Foundry tests against Madara RPC